+++ lib/enums.dart
@@ -1,0 +1,16 @@
+enum StylishBarFabLocation { center, end }
+
+enum BarAnimation {
+  fade,
+  blink,
+  transform3D,
+}
+
+enum IconStyle {
+  simple,
+  animated,
+}
+
+enum BubbleBarStyle { vertical, horizotnal }
+
+enum BubbleFillStyle { fill, outlined }